created a seperate test for the GenomeGraph class
